pyproject.toml
[project]
name = "voidrp-account-backend"
version = "0.1.0"
description = "VoidRP account-core backend foundation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.15.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.0.0",
  "PyJWT>=2.8.0",
  "pwdlib[argon2]>=0.3.0",
  "email-validator>=2.0.0",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.28.0",
]

[tool.pytest.ini_options]
pythonpath = [
  ".",
]
testpaths = [
  "tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.setuptools]
package-dir = {"" = "apps/api"}

[tool.setuptools.packages.find]
where = ["apps/api"]
include = ["app*"]
exclude = ["tests*", "alembic*"]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"